Re-organise endpoints organisation
